Move encoding test
